requirements.txt
aio-pika==6.4.1
aioelasticsearch==0.6.0
aiohttp==3.6.2
aiohttp-jwt==0.5.0
aiomysql==0.0.20
aioredis==1.3.1
aiormq==3.2.0
async-timeout==3.0.1
attrs==19.3.0
automium==0.2.6
automium-web==0.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
elasticsearch==7.1.0
Flask==1.1.1
hiredis==1.0.1
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.3.0
inflect==4.0.0
itsdangerous==1.1.0
Jinja2==2.11.1
legacy==0.1.6
MarkupSafe==1.1.1
more-itertools==8.0.2
multidict==4.7.3
mysql-replication==0.21
pamqp==2.3.0
pika==1.1.0
public==2019.4.13
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pynsq==0.9.0b1
query-string==2019.4.13
redis==3.3.11
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.12
tornado==5.1.1
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.4.2
zipp==0.6.0
emoji==0.5.4